services/backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   -r requirements.in
    #   django-celery-results
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
commonmark==0.9.1
    # via rich
cryptography==38.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.13
    # via redis
django==4.1.1
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-debug-toolbar
    #   django-filter
    #   django-model-utils
    #   django-modeltranslation
    #   djangorestframework
django-allauth==0.51.0
    # via -r requirements.in
django-celery-results==2.4.0
    # via -r requirements.in
django-debug-toolbar==3.7.0
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-modeltranslation==0.18.4
    # via -r requirements.in
django-widget-tweaks==1.4.12
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
kombu==5.2.4
    # via celery
markdown==3.4.1
    # via -r requirements.in
oauthlib==3.2.1
    # via requests-oauthlib
packaging==21.3
    # via redis
prompt-toolkit==3.0.31
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via rich
pyjwt[crypto]==2.5.0
    # via django-allauth
pyparsing==3.0.9
    # via packaging
python3-openid==3.2.0
    # via django-allauth
pytz==2022.2.1
    # via
    #   celery
    #   djangorestframework
redis==4.3.4
    # via -r requirements.in
requests==2.28.1
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
rich==12.5.1
    # via -r requirements.in
six==1.16.0
    # via click-repl
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
types-cryptography==3.3.23
    # via pyjwt
tzdata==2022.4
    # via -r requirements.in
urllib3==1.26.12
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.14.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
